+++ rxStore/src/test/kotlin/com/dropbox/store/rx/InMemoryRxPersister.kt
@@ -10,11 +10,11 @@ class InMemoryRxPersister<Key, Output> {
     private val data = mutableMapOf<Key, Output>()
 
     @Suppress("RedundantSuspendModifier") // for function reference
-     fun read(key: Key): Maybe<Output> = if(data[key]!=null) Maybe.just(data[key]) else Maybe.empty()
+     fun read(key: Key): Maybe<Output> {
+        return Maybe.fromCallable {data[key]}
+     }
 
     @Suppress("RedundantSuspendModifier") // for function reference
-     fun write(key: Key, output: Output): Single<Unit> {
-        data[key] = output
-        return Single.just(Unit)
-    }
+     fun write(key: Key, output: Output): Single<Unit> =
+        Single.fromCallable { data[key] = output }
 }
