+++ store/src/test/java/com/dropbox/android/external/store4/impl/FlowStoreTest.kt
@@ -108,6 +108,7 @@ class FlowStoreTest {
             3 to "three-2"
         )
         val persister = InMemoryPersister<Int, String>()
+
         val pipeline = build(
             nonFlowingFetcher = fetcher::fetch,
             persisterReader = persister::read,
